Create childtheme-guide
This is a guide to starting a child theme on UCONN latest theme. In the 2014 school year Web Development Lab of UITS will be releasing a new theme called Cornerstone. It will be highly customizable and be easy for users to create and modify their website. For those users who would like to make their website even more unique this guide will help them start out.